pyproject.toml
[project]
name = "voice-live-agent"
version = "0.1.0"
description = "Real-time voice agent with Azure Voice Live and ACS/Web streaming clients"
dependencies = [
    "quart",
    "websockets",
    "httpx",
    "python-dotenv",
    "quart-cors",
    "azure-core",
    "azure-communication-identity",
    "aiohttp",
    "azure-eventgrid",
    "azure-identity",
    "azure-communication-callautomation",
    "openai[realtime]",
    "azure-communication-email",
    "aiomysql>=0.2.0",
]
requires-python = ">=3.9"
